Refactor hiring module
- upgrade to directory
- introduce application, opening, staking_policy submodules
- copy related code to the new modules
- implement several macro’s as generic functions
